Create exponentialOrNoisyMaxDP.py
Differential privacy is implemented in this branch using an entirely different method. Rather than adding noise directly to returned values, we use an exponential implementation on classes. Also provided is a method that adds noise and reports the maximum noisy value for all classes. This can be used in place of the exponential method when the classes are finite, which in the case of the project as it currently stands, is.